build/core: Update virtual memory map to move mmap to 0x2AAAAAAA
mmap start depends on kernel configuration.  This mmap address is equal to
PAGE_OFFSET_BASE/3.  With kernel 2G/2G split for Android targets, this
mmap is being updated accordingly.  If 1G/3G split is used, the prelink
map needs to be updated accordingly.  The kernel does not respect this
prelink map, but instead Android prelinking is done around kernel config.
This means, if the VM split of kernel changes, this needs to change as
well.  Also moved thread stack closer to code/data as it originally was
and moved prelinked and non prelinked libraries above mmap'ed area.