Fix for Intent droiddoc failure
In some infrequent cases, the two uses of indexOf in isHidden appear to return
true to indicate that @hide or @pending is present in the doc string for Intent
when this is not the case.  This change replaces the use of indexOf with a
custom coded verson which handles less cases and proves more reliable under
test.

Change-Id: I0cfd37b00e320e1fe6aa39bbe443e362d36aaf1a