Modify droiddoc generation to avoid find errors
Change the code in droiddoc generation to avoid errors from find.
The SDK and target builds hand different paths to droiddoc: one
with LOCAL_PATH set, the other without.  droiddoc then prepends
LOCAL_PATH.  This change checks to make sure it's not already
there before prepending.
The second find error calls find on a directory that may be
non-existent.  This change doesn't invoke find on the directory
if the directory does not exist.

Change-Id: I539524106c85b23460597dd273ea5690622093ef