Fix get-kernel-size macro
In the 3.4 kernel the _text symbol is used to represent the
beginning of the kernel image. Note that a space needs to
be present before _text in the search string so that other
symbols with names ending in _text are not picked up.

(cherry picked from commit 1ccee9dbb1a1c769ea9d9619ee931072d25cf207)

Change-Id: I1cfcd8526195cf44faedc11174e5fb29a9475286