Bug 916863: Added NFC configuration
On the emulator, nfcd and nfc_nci.generic have to be build. Also
the variable TARGET_BOOTLOADER_BOARD_NAME has to be set to the
value 'generic', so that libhal will find this value in the env
property 'ro.product.board' when it loads nfc_nci.generic. And NFC
has to be enabled by setting the ro.moz.nfc.enabled.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>