Bug 1004195: Handle bootable/recovery/res being a symlink.
Some versions of `bootable/recovery` replace the subdirectory `res` with
a symlink to a copy in `librecovery`, so that FxOS-specific graphics can
be used from the mozilla-b2g fork of that project.  However, this causes
`cp -rf` to try to copy the symlink itself (and fail), and `find` to not
resolve the symlink to find the actual dependencies.

The trailing slash added in this patch fixes these problems.