build/tools: Fix problem with unicode file names
Unicode file names can be returned through filesystem either as
UTF-8 or raw (i.e. escaped Unicode Octets). This can cause errors
as we are using the filenames as keys in a dict. The lookup can
incorrectly fail in the dict if key was stored as non-UTF8 and then
looked up as UTF-8.

This change always stores and retrieves the dict keys as UTF-8 to
resolve the problem.

Change-Id: I5b4de7104dea5de33f1e9a76a1e0087bcdf246d8
CRs-Fixed: 1068027