Add recovery & recoveryfs upgrade support
target-file*.zip has been packed with a recovery.img
(which is essentially the same as boot.img) and an unsparsed
recoveryfs image as well. Add support to include the 2 images
as part of update*.zip in a new directory - recoveryupgrade.
Both images are packed whole and not incrementally as patches.

We currently only support recovery-upgrade for emmc devices.

For recovery images to be packed into update*.zip,
recovery_upgrade_supported should be "1" in META/misc_info.txt.

Change-Id: Ib66c6db9a0424b338e0b811453eb6a498a9f5aa4