APQ8053: Add support for recoverypatch.
Comment out the patching of recovery-from-boot.p as of now.
By commenting it we are disabling patching of the recovery partition.
Need to still create the recovery-from-boot.p as block based flow is
very tightly dependent on recovery-from-boot.p hence did not disrupted
that.

Also after OTA is done there will be a hanging recovery-from-boot.p
saved on the system partition. As there is no init script in LE to rm
it.

Can enable it if there is any requirement later.

Change-Id: I1d5e419c8b0bdfae0633c54242ed0f3387b931cc