Copy dex file to javalib.jar for prebuilt JAVA_LIBRARIES.
Copy dex file to javalib.jar for prebuilt JAVA_LIBRARIES.
Right now classes.jar is copied to javalib.jar for prebuilts.
It causes an issue for PRODUCT_BOOT_JARS because dex_bootjars
expects javalib.jar to have dex file.But it doesn't have
classes.dex in it.
Add LOCAL_DEX_FILE variable to define the jar with dex to be copied.

Change-Id: If5bd56df0d5c54d70ad20d4cebebd6507ce3f34c