Support non-standard partition names
This change makes it possible to configure the boot, system and vendor
partition names in a BoardConfig makefile.
This is needed when AVB is used with Android as a guest VM. There, the
partition names may not necessarily be "boot", "system" and "vendor".
For the "system" and "vendor" partitions, allow the partition name to be
different, while mount point is still "system" or "vendor".
Additionally, for the boot partition, add an option in Boardconfig to
remove AVB footer after boot image is generated and VBMeta data was
collected from the image.
This is since when VBMeta partition is used, the AVB footer is not used,
and it dramatically increases the boot image size.

Change-Id: Ib471f23a64409e12522f606a1d04979c75bbc36d
Signed-off-by: Amit Blay <ablay@codeaurora.org>